Dialogs/Waypoint/WaypointInfoWidget: add reciprocal RWY name
Calculate the reciprocal runway direction and display its name (number) next to the main runway. As most runways can normally be used in both directions it saves the pilot some mental effort to caclulate the opposite direction in his/her head when needed. An explicit note in the waypoint description is a way to inform if a runway is one-directional only.